dental_crm_python_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
mysql-connector-python==8.3.0
openai==1.10.0




